+++ dataloom/statements/__init__.py
@@ -229,35 +229,62 @@ class GetStatement[T]:
             "" if offset is None else f"OFFSET { offset}",
         ]
         if self.dialect == "postgres":
-            sql = PgStatements.SELECT_WHERE_COMMAND.format(
-                column_names=", ".join(fields)
-                if function is not None
-                else ", ".join([f'"{f}"' for f in fields] + group_fns),
-                table_name=f'"{self.table_name}"',
-                filters=" ".join(placeholder_filters),
-                options=" ".join(options),
-                distinct="DISTINCT" if distinct else "",
-            )
+            if function is None:
+                sql = PgStatements.SELECT_WHERE_COMMAND.format(
+                    column_names=", ".join([f'"{f}"' for f in fields] + group_fns),
+                    table_name=f'"{self.table_name}"',
+                    filters=" ".join(placeholder_filters),
+                    options=" ".join(options),
+                    distinct="DISTINCT" if distinct else "",
+                )
+            else:
+                sql = PgStatements.SELECT_WHERE_FN_COMMAND.format(
+                    fn=function.upper(),
+                    column_names=", ".join(fields),
+                    table_name=f'"{self.table_name}"',
+                    filters=" ".join(placeholder_filters),
+                    options=" ".join(options),
+                )
         elif self.dialect == "mysql":
-            sql = MySqlStatements.SELECT_WHERE_COMMAND.format(
-                column_names=", ".join(fields)
-                if function is not None
-                else ", ".join([f"`{name}`" for name in fields] + group_fns),
-                table_name=f"`{self.table_name}`",
-                filters=" ".join(placeholder_filters),
-                options=" ".join(options),
-                distinct="DISTINCT" if distinct else "",
-            )
+            if function is None:
+                sql = MySqlStatements.SELECT_WHERE_COMMAND.format(
+                    column_names=", ".join(
+                        [f"`{name}`" for name in fields] + group_fns
+                    ),
+                    table_name=f"`{self.table_name}`",
+                    filters=" ".join(placeholder_filters),
+                    options=" ".join(options),
+                    distinct="DISTINCT" if distinct else "",
+                )
+            else:
+                sql = MySqlStatements.SELECT_WHERE_FN_COMMAND.format(
+                    fn=function.upper(),
+                    column_names=", ".join(fields),
+                    table_name=f"`{self.table_name}`",
+                    filters=" ".join(placeholder_filters),
+                    options=" ".join(options),
+                )
+
         elif self.dialect == "sqlite":
-            sql = Sqlite3Statements.SELECT_WHERE_COMMAND.format(
-                column_names=", ".join(fields)
-                if function is not None
-                else ", ".join([f"`{name}`" for name in fields] + group_fns),
-                table_name=f"`{self.table_name}`",
-                filters=" ".join(placeholder_filters),
-                options=" ".join(options),
-                distinct="DISTINCT" if distinct else "",
-            )
+            if function is None:
+                sql = Sqlite3Statements.SELECT_WHERE_COMMAND.format(
+                    column_names=", ".join(
+                        [f"`{name}`" for name in fields] + group_fns
+                    ),
+                    table_name=f"`{self.table_name}`",
+                    filters=" ".join(placeholder_filters),
+                    options=" ".join(options),
+                    distinct="DISTINCT" if distinct else "",
+                )
+            else:
+                sql = Sqlite3Statements.SELECT_WHERE_FN_COMMAND.format(
+                    fn=function.upper(),
+                    column_names=", ".join(fields),
+                    table_name=f"`{self.table_name}`",
+                    filters=" ".join(placeholder_filters),
+                    options=" ".join(options),
+                )
+
         else:
             raise UnsupportedDialectException(
                 "The dialect passed is not supported the supported dialects are: {'postgres', 'mysql', 'sqlite'}"
@@ -286,32 +313,57 @@ class GetStatement[T]:
         ]
 
         if self.dialect == "postgres":
-            sql = PgStatements.SELECT_COMMAND.format(
-                column_names=", ".join(fields)
-                if function is not None
-                else ", ".join([f'"{name}"' for name in fields] + group_fns),
-                table_name=f'"{self.table_name}"',
-                options=" ".join(options),
-                distinct="DISTINCT" if distinct else "",
-            )
+            if function is None:
+                sql = PgStatements.SELECT_COMMAND.format(
+                    column_names=", ".join(
+                        [f'"{name}"' for name in fields] + group_fns
+                    ),
+                    table_name=f'"{self.table_name}"',
+                    options=" ".join(options),
+                    distinct="DISTINCT" if distinct else "",
+                )
+            else:
+                sql = PgStatements.SELECT_FN_COMMAND.format(
+                    fn=function.upper(),
+                    column_names=", ".join(fields),
+                    table_name=f'"{self.table_name}"',
+                    options=" ".join(options),
+                )
         elif self.dialect == "mysql":
-            sql = MySqlStatements.SELECT_COMMAND.format(
-                column_names=", ".join(fields)
-                if function is not None
-                else ", ".join([f"`{name}`" for name in fields] + group_fns),
-                table_name=f"`{self.table_name}`",
-                options=" ".join(options),
-                distinct="DISTINCT" if distinct else "",
-            )
+            if function is None:
+                sql = MySqlStatements.SELECT_COMMAND.format(
+                    column_names=", ".join(fields)
+                    if function is not None
+                    else ", ".join([f"`{name}`" for name in fields] + group_fns),
+                    table_name=f"`{self.table_name}`",
+                    options=" ".join(options),
+                    distinct="DISTINCT" if distinct else "",
+                )
+            else:
+                sql = MySqlStatements.SELECT_FN_COMMAND.format(
+                    fn=function.upper(),
+                    column_names=", ".join(fields),
+                    table_name=f"`{self.table_name}`",
+                    options=" ".join(options),
+                )
+
         elif self.dialect == "sqlite":
-            sql = Sqlite3Statements.SELECT_COMMAND.format(
-                column_names=", ".join(fields)
-                if function is not None
-                else ", ".join([f"`{name}`" for name in fields] + group_fns),
-                table_name=f"`{self.table_name}`",
-                options=" ".join(options),
-                distinct="DISTINCT" if distinct else "",
-            )
+            if function is None:
+                sql = Sqlite3Statements.SELECT_COMMAND.format(
+                    column_names=", ".join(
+                        [f"`{name}`" for name in fields] + group_fns
+                    ),
+                    table_name=f"`{self.table_name}`",
+                    options=" ".join(options),
+                    distinct="DISTINCT" if distinct else "",
+                )
+            else:
+                sql = Sqlite3Statements.SELECT_FN_COMMAND.format(
+                    fn=function.upper(),
+                    column_names=", ".join(fields),
+                    table_name=f"`{self.table_name}`",
+                    options=" ".join(options),
+                )
         else:
             raise UnsupportedDialectException(
                 "The dialect passed is not supported the supported dialects are: {'postgres', 'mysql', 'sqlite'}"