+++ dataloom/statements/statements.py
@@ -92,6 +92,21 @@ class MySqlStatements:
     )
     SELECT_BY_PK = "SELECT {column_names} FROM {table_name} WHERE {pk_name} = {pk};"
     SELECT_WHERE_COMMAND = "SELECT {distinct} {column_names} FROM {table_name} WHERE {filters} {options};".strip()
+    SELECT_WHERE_FN_COMMAND = """
+    SELECT {fn}({column_names}) FROM (
+            SELECT {column_names}
+            FROM {table_name}
+            WHERE {filters}
+            {options}
+    ) AS subquery;
+    """
+    SELECT_FN_COMMAND = """
+    SELECT {fn}({column_names}) FROM (
+            SELECT {column_names}
+            FROM {table_name}
+            {options}
+    ) AS subquery;
+    """
 
     # ------------- child parent bidirectional sub queries
     SELECT_CHILD_BY_PK = """
@@ -219,6 +234,21 @@ class Sqlite3Statements:
     )
     SELECT_BY_PK = "SELECT {column_names} FROM {table_name} WHERE {pk_name} = {pk};"
     SELECT_WHERE_COMMAND = "SELECT {distinct} {column_names} FROM {table_name} WHERE {filters} {options};".strip()
+    SELECT_WHERE_FN_COMMAND = """
+    SELECT {fn}({column_names}) FROM (
+            SELECT {column_names}
+            FROM {table_name}
+            WHERE {filters}
+            {options}
+    ) AS subquery;
+    """
+    SELECT_FN_COMMAND = """
+    SELECT {fn}({column_names}) FROM (
+            SELECT {column_names}
+            FROM {table_name}
+            {options}
+    ) AS subquery;
+    """
 
     # ------------- child parent bidirectional sub queries
     SELECT_CHILD_BY_PK = """
@@ -344,6 +374,21 @@ class PgStatements:
     )
     SELECT_BY_PK = "SELECT {column_names} FROM {table_name} WHERE {pk_name} = {pk};"
     SELECT_WHERE_COMMAND = "SELECT {distinct} {column_names} FROM {table_name} WHERE {filters} {options};".strip()
+    SELECT_WHERE_FN_COMMAND = """
+    SELECT {fn}({column_names}) FROM (
+            SELECT {column_names}
+            FROM {table_name}
+            WHERE {filters}
+            {options}
+    ) AS subquery;
+    """
+    SELECT_FN_COMMAND = """
+    SELECT {fn}({column_names}) FROM (
+            SELECT {column_names}
+            FROM {table_name}
+            {options}
+    ) AS subquery;
+    """
 
     # ------------- child parent bidirectional sub queries
     SELECT_CHILD_BY_PK = """